pyproject.toml
[tool.poetry]
name = "online-shop"
version = "0.1.0"
description = ""
authors = ["pedram"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"
pillow = "^10.3.0"
selenium = "^4.20.0"
redis = "^5.0.4"
kavenegar = "^1.1.2"
pycryptodome = "^3.20.0"
celery = "^5.4.0"
django-storages = "^1.14.3"
django-celery-beat = "^2.6.0"
boto3 = "^1.34.103"
django-celery-results = "^2.5.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
ruff = "^0.4.3"
pre-commit = "^3.7.0"
model-bakery = "^1.18.0"
gunicorn = "^22.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
